[SQOOP-2839]
Fix For SQOOP-2839. On our investigation we found out this is related to Sqoop’s internal implementation. Sqoop internally creates a Java class corresponding to the table to be imported, which contains all the column names as data members. It also includes a constant named  "PROTOCOL_VERSION" and this leads to conflict between data members, if one of the column is named “PROTOCOL_VERSION”.

So in order to avoid this situation , adding "PROTOCOL_VERSION" in reserved keywords