[SQOOP-2846]
After analyzing the issue, it is found that export with update key is
not implemneted properly in case of avro and parqute file formats.

In order to correct the issue [SQOOP-2846], export with update option is
implemented for the mentioned file formats. For this appropriate
mapper, input format and output format classes were used in the
implementation.

Updated test cases were executed successfully and the same is also
incorporated in the patch.